// src/app/models/invoice.ts
export interface Invoice {
    FECHA_REGISTRO?: Date;
    CODIGO_FACTURA?: number;
    CIUDAD_CLIENTE?: string;
    TIPO_CLIENTE?: string;
    ID_CLIENTE?: number;
    NOMBRES_CLIENTE?: string;
    DIRECCION_CLIENTE?: string;
    TELEFONO_CLIENTE?: string;
    CORREO_CLIENTE?: string;
    ID_ALUMNO?: number;
    NOMBRES_ALUMNO?: string;
    APELLIDOS_ALUMNO?: string;
    GRADO_ALUMNO?: string;
    CONCEPTO_1?: string;
    MONTO_1_USD?: number;
    MONTO_1_BS?: number;
    MONTO_1_BCV?: number;
    CONCEPTO_2?: string;
    MONTO_2_USD?: number;
    MONTO_2_BS?: number;
    MONTO_2_BCV?: number;
    CONCEPTO_3?: string;
    MONTO_3_USD?: number;
    MONTO_3_BS?: number;
    MONTO_3_BCV?: number;
    CONCEPTO_4?: string;
    MONTO_4_USD?: number;
    MONTO_4_BS?: number;
    MONTO_4_BCV?: number;
    CONCEPTO_5?: string;
    MONTO_5_USD?: number;
    MONTO_5_BS?: number;
    MONTO_5_BCV?: number;
    TIPO_TRANSACCION?: string;
    NO_TRANSACCION?: string;
    EFECTIVO?: number;
    BANCO?: string;
    SUBTOTAL?: number;
    IVA?: number;
    TOTAL?: number;
    CREDITO_CONTADO?: string;
    OBSERVACIONES?: string;
    REGISTRADO_POR?: string;
    ESTATUS?: boolean;
    ADVERTENCIA?: string;
    MOD_FECHA_REGISTRO?: Date;
    }